openpower-dumps: Move opdump utility functions to utility file
